+++ tests/test_define.h
@@ -45,6 +45,8 @@ extern "C" {
 #define TEST_NON_TRYTE_TAG "tangleaccelerator9999999999"
 #define VALUE 100
 #define TIMESTAMP 1546436542
+#define TIMESTAMP_LEN20_1 00000000001546436542
+#define TIMESTAMP_LEN20_2 00000000001546438871
 #define CURRENT_INDEX 1
 #define LAST_INDEX 2
 #define NONCE "THISTANGLEACCELERATORNONCES"
